Revert "Change type of val length to size_t"
This reverts commit 9d30ddbba73b0a1b6d141e0717b2c2fe90aca67f.

+++ src/aof.c
@@ -1096,7 +1096,7 @@ int rewriteSortedSetObject(rio *r, robj *key, robj *o) {
         unsigned char *zl = o->ptr;
         unsigned char *eptr, *sptr;
         unsigned char *vstr;
-        size_t vlen;
+        unsigned int vlen;
         long long vll;
         double score;
 
@@ -1176,7 +1176,7 @@ int rewriteSortedSetObject(rio *r, robj *key, robj *o) {
 static int rioWriteHashIteratorCursor(rio *r, hashTypeIterator *hi, int what) {
     if (IS_PACKED(hi)) {
         unsigned char *vstr = NULL;
-        size_t vlen = UINT_MAX;
+        unsigned int vlen = UINT_MAX;
         long long vll = LLONG_MAX;
 
         hashTypeCurrentFromZiplist(hi, what, &vstr, &vlen, &vll);
