+++ src/ziplist.h
@@ -39,19 +39,19 @@
 
 unsigned char *ziplistNew(void);
 unsigned char *ziplistMerge(unsigned char **first, unsigned char **second);
-unsigned char *ziplistPush(unsigned char *zl, unsigned char *s, size_t slen, int where);
-unsigned char *ziplistPushHead(unsigned char *zl, unsigned char *s, size_t slen);
-unsigned char *ziplistPushTail(unsigned char *zl, unsigned char *s, size_t slen);
+unsigned char *ziplistPush(unsigned char *zl, unsigned char *s, unsigned int slen, int where);
+unsigned char *ziplistPushHead(unsigned char *zl, unsigned char *s, unsigned int slen);
+unsigned char *ziplistPushTail(unsigned char *zl, unsigned char *s, unsigned int slen);
 unsigned char *ziplistIndex(unsigned char *zl, long index);
 unsigned char *ziplistNext(unsigned char *zl, unsigned char *p);
 unsigned char *ziplistPrev(unsigned char *zl, unsigned char *p);
-unsigned int ziplistGet(unsigned char *p, unsigned char **sval, size_t *slen, long long *lval);
-unsigned char *ziplistInsert(unsigned char *zl, unsigned char *p, unsigned char *s, size_t slen);
+unsigned int ziplistGet(unsigned char *p, unsigned char **sval, unsigned int *slen, long long *lval);
+unsigned char *ziplistInsert(unsigned char *zl, unsigned char *p, unsigned char *s, unsigned int slen);
 unsigned char *ziplistDelete(unsigned char *zl, unsigned char **p);
 unsigned char *ziplistDeleteRange(unsigned char *zl, int index, unsigned int num);
-unsigned char *ziplistReplace(unsigned char *zl, unsigned char *p, unsigned char *s, size_t slen);
-unsigned int ziplistCompare(unsigned char *p, unsigned char *s, size_t slen);
-unsigned char *ziplistFind(unsigned char *zl, unsigned char *p, unsigned char *vstr, size_t vlen, unsigned int skip);
+unsigned char *ziplistReplace(unsigned char *zl, unsigned char *p, unsigned char *s, unsigned int slen);
+unsigned int ziplistCompare(unsigned char *p, unsigned char *s, unsigned int slen);
+unsigned char *ziplistFind(unsigned char *zl, unsigned char *p, unsigned char *vstr, unsigned int vlen, unsigned int skip);
 long ziplistLen(unsigned char *zl);
 size_t ziplistBlobLen(unsigned char *zl);
 void ziplistRepr(unsigned char *zl);
